+++ HISTORY.md
@@ -3,6 +3,8 @@
 
 ## not yet released, version 1.2.1-SNAPSHOT
 
+- Fixed missing parentheses in the `toTex` representation of function
+  `permutations`.
 - Some minor fixes in the docs.